replace Float to Long values

+++ playkit/src/main/java/com/kaltura/playkit/ads/PKAdInfo.java
@@ -15,5 +15,5 @@ public interface PKAdInfo {
     int      getAdPodPosition();
     long     getAdPodTimeOffset();
     long     getAdDuration();
-    List<Float> getAdCuePoints();
+    List<Long> getAdCuePoints();
 }
